requirements.txt
black==19.10b0
pandas==1.1.4
python-dotenv==0.14.0
